moving ethclient to platform/blockchain
