+++ internal/platform/blockchain/client.go
@@ -1,0 +1,27 @@
+package blockchain
+
+import (
+	"context"
+	"math/big"
+
+	"github.com/ethereum/go-ethereum"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+)
+
+// EthClient is an interface for interacting with Ethereum.
+// https://github.com/ethereum/go-ethereum/pull/23884
+type EthClient interface {
+	ethereum.ChainReader
+	ethereum.TransactionReader
+	ethereum.ChainSyncReader
+	ethereum.ContractCaller
+	ethereum.LogFilterer
+	ethereum.TransactionSender
+	ethereum.GasPricer
+	ethereum.PendingContractCaller
+	ethereum.GasEstimator
+	bind.ContractBackend
+	ChainID(ctx context.Context) (*big.Int, error)
+	BlockNumber(ctx context.Context) (uint64, error)
+	Close()
+}
